added result count service

+++ app/models/db/attempts.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, Time
+from sqlalchemy import DECIMAL, Column, DateTime, ForeignKey, Integer, Time
 from sqlalchemy.orm import relationship
 
 from app.core.database import Base
@@ -17,7 +17,7 @@ class Attempt(Base):
     start_time = Column(DateTime, default=datetime.utcnow())
     end_time = Column(DateTime)
     spent_time = Column(Time)
-    result = Column(Integer, default=0)
+    result = Column(DECIMAL, default=0)
 
     def __repr__(self) -> str:
         return f"Attempt for quiz {self.quiz_id}"
